+++ src/main/java/org/kingpixel/cobblemonpatches/mixins/cobblemon/FormDataMixin.java
@@ -1,6 +1,6 @@
 package org.kingpixel.cobblemonpatches.mixins.cobblemon;
 
-import com.cobblemon.mod.common.pokemon.Species;
+import com.cobblemon.mod.common.pokemon.FormData;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -8,12 +8,13 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 /**
- * @author Carlos Varas Alonso - 27/10/2025 6:05
+ *
+ * @author Carlos Varas Alonso - 26/07/2026 19:13
  */
-@Mixin(value = Species.class, remap = false)
-public abstract class SpeciesMixin {
+@Mixin(value = FormData.class, remap = false)
+public abstract class FormDataMixin {
 
-  // Improve performance of showdownId() by caching the result
+  @Unique private String formOnlyShowdownIdCache = null;
   @Unique private String showdownIdCache = null;
 
   @Inject(method = "showdownId", at = @At("HEAD"), cancellable = true)
@@ -27,6 +28,20 @@ public abstract class SpeciesMixin {
   private void cobblemonPatches$returnShowdownId(CallbackInfoReturnable<String> cir) {
     if (showdownIdCache == null && cir.getReturnValue() != null) {
       showdownIdCache = cir.getReturnValue().intern();
+    }
+  }
+
+  @Inject(method = "formOnlyShowdownId", at = @At("HEAD"), cancellable = true)
+  private void cobblemonPatches$headFormOnlyShowdownId(CallbackInfoReturnable<String> cir) {
+    if (formOnlyShowdownIdCache != null) {
+      cir.setReturnValue(formOnlyShowdownIdCache);
+    }
+  }
+
+  @Inject(method = "formOnlyShowdownId", at = @At("RETURN"))
+  private void cobblemonPatches$returnFormOnlyShowdownId(CallbackInfoReturnable<String> cir) {
+    if (formOnlyShowdownIdCache == null && cir.getReturnValue() != null) {
+      formOnlyShowdownIdCache = cir.getReturnValue().intern();
     }
   }
 }